refactor: useFactory 함수 분리
