+++ server/src/common/redis/redis.module.ts
@@ -1,6 +1,6 @@
 import { Global, Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { createClient, RedisClientType } from 'redis';
+import { createRedisProvider } from './redis.provider';
 
 export const REDIS_CLIENT = 'REDIS_CLIENT' as const;
 
@@ -10,20 +10,7 @@ export const REDIS_CLIENT = 'REDIS_CLIENT' as const;
   providers: [
     {
       provide: REDIS_CLIENT,
-      useFactory: async (
-        configService: ConfigService,
-      ): Promise<RedisClientType> => {
-        const client = createClient({
-          socket: {
-            host: configService.get('REDIS_HOST'),
-            port: configService.get('REDIS_PORT'),
-          },
-          password: configService.get('REDIS_PASSWORD'),
-        }) as RedisClientType;
-
-        await client.connect();
-        return client;
-      },
+      useFactory: createRedisProvider,
       inject: [ConfigService],
     },
   ],
